move editor hook
